Delete backend directory
Don't forget we are the lions. The real rulers. This is a fact..we have been giving a whoop ass to your oppression and evil forces...king Solomon. Prophet Muhammad..ranseys the third..another fact...stolen valour stolen valour. U have something that's isn't yours, it belongs to me. Hello?